Register service worker
This commit also removes hashes from all webpack bundle filenames.
Service workers' filename must be stable rather than include a
a cache-busting hash or the browser doesn't treat it as the same
SW. While it's possible to configure different webpack entry points
to have different output filenames, this appeared difficult to do
within webpacker, so we elected to defer that work and maintain
velocity. We'll want to restore filename hashes to application.js
(though NOT to servie-worker.js) prior to production use.