include request method, headers, and body in indexeddb
In order to replay a deferred request to add a record to the database, we need
at least the url and method (to dispatch to the correct controller action), the
headers (to satisfy authentication requirements), and the body. Before this
change we only stored the url in indexeddb, and after this change we will also
store the method, headers, and body.