Changed column from percent_filled to corals_added
On our last call with the Coral Gardeners, they clarified that they do not want to log the percent of a table filled when they are on a dive. They merely want to log the number added and removed.
This pr just does the number added because someone else is doing number_removed
TODO: Translate for i18n